requirements.txt
aiogram==3.*
openai>=1.0.0
python-dotenv
pillow
aiosqlite

